Move Resources to services/resources-impl.js
