+++ src/amp-core-service.js
@@ -16,6 +16,7 @@
 
 import {installActionService} from './service/action-impl';
 import {installHistoryService} from './service/history-impl';
+import {installResourcesService} from './service/resources-impl';
 import {installViewerService} from './service/viewer-impl';
 import {installViewportService} from './service/viewport-impl';
 import {installVsyncService} from './service/vsync-impl';
@@ -32,4 +33,5 @@ export function installCoreServices(window) {
   installHistoryService(window);
   installVsyncService(window);
   installActionService(window);
+  installResourcesService(window);
 }
